+++ packages/zowe-explorer-api/__tests__/__unit__/profiles/ZoweExplorerZosmfApi.unit.test.ts
@@ -896,7 +896,7 @@ describe("ZosmfCommandApi", () => {
             transform: () => ["command", { addressSpaceOptions: { account: "ACCT#" } }],
         },
         {
-            name: "issueMvsCommand",
+            name: "issueConsoleCommand",
             spy: sharedSpyOn(zosconsole.IssueCommand, "issue"),
             args: ["command", "defcn"],
             transform: (args) => [{ command: args[0], consoleName: args[1], processResponses: true }],