+++ packages/zowe-explorer-api/src/extend/MainframeInteraction.ts
@@ -553,13 +553,13 @@ export namespace MainframeInteraction {
         issueTsoCommandWithParms?(command: string, parms?: zostso.IStartTsoParms, useNewTsoApis?: boolean): Promise<zostso.IIssueResponse>;
 
         /**
-         * Issues a MVS Command and returns a Console Command API response.
+         * Issues a console Command and returns a Console Command API response.
          *
          * @param {string} command
          * @returns {Promise<zosconsole.IConsoleResponse>}
          * @memberof ICommand
          */
-        issueMvsCommand?(command: string, consoleName?: string): Promise<zosconsole.IConsoleResponse>;
+        issueConsoleCommand?(command: string, consoleName?: string): Promise<zosconsole.IConsoleResponse>;
 
         /**
          * Issues a Unix Command and returns a Console Command API response.