+++ packages/zowe-explorer-api/src/profiles/ZoweExplorerZosmfApi.ts
@@ -530,7 +530,7 @@ import { IDataSetCount } from "../dataset/IDataSetCount";
             return zostso.IssueTso.issueTsoCommand(this.getSession(), parms.account, command, parms);
         }
 
-        public issueMvsCommand(command: string, consoleName?: string): Promise<zosconsole.IConsoleResponse> {
+        public issueConsoleCommand(command: string, consoleName?: string): Promise<zosconsole.IConsoleResponse> {
             return zosconsole.IssueCommand.issue(this.getSession(), { command, consoleName, processResponses: true });
         }