+++ packages/zowe-explorer/__tests__/__unit__/trees/shared/SharedInit.unit.test.ts
@@ -22,7 +22,7 @@ import { LocalFileManagement } from "../../../../src/management/LocalFileManagem
 import { ZoweExplorerApiRegister } from "../../../../src/extending/ZoweExplorerApiRegister";
 import { SharedInit } from "../../../../src/trees/shared/SharedInit";
 import { TsoCommandHandler } from "../../../../src/commands/TsoCommandHandler";
-import { MvsCommandHandler } from "../../../../src/commands/MvsCommandHandler";
+import { ConsoleCommandHandler } from "../../../../src/commands/ConsoleCommandHandler";
 import { UnixCommandHandler } from "../../../../src/commands/UnixCommandHandler";
 import { SharedTreeProviders } from "../../../../src/trees/shared/SharedTreeProviders";
 import { SharedContext } from "../../../../src/trees/shared/SharedContext";
@@ -63,7 +63,7 @@ describe("Test src/shared/extension", () => {
             _: { _: "_" },
         };
         const profileMocks = { deleteProfile: vi.fn(), disableValidation: vi.fn(), enableValidation: vi.fn(), refresh: vi.fn() };
-        const cmdProviders = { mvs: { issueMvsCommand: vi.fn() }, tso: { issueTsoCommand: vi.fn() }, uss: { issueUnixCommand: vi.fn() } };
+        const cmdProviders = { mvs: { issueConsoleCommand: vi.fn() }, tso: { issueTsoCommand: vi.fn() }, uss: { issueUnixCommand: vi.fn() } };
         const treeProvider = {
             addFavorite: vi.fn(),
             deleteSession: vi.fn(),
@@ -332,12 +332,12 @@ describe("Test src/shared/extension", () => {
             },
             {
                 name: "zowe.issueMvsCmd:1",
-                mock: [{ spy: vi.spyOn(cmdProviders.mvs, "issueMvsCommand"), arg: [undefined, undefined, test.value] }],
+                mock: [{ spy: vi.spyOn(cmdProviders.mvs, "issueConsoleCommand"), arg: [undefined, undefined, test.value] }],
             },
             {
                 name: "zowe.issueMvsCmd:2",
                 parm: [],
-                mock: [{ spy: vi.spyOn(cmdProviders.mvs, "issueMvsCommand"), arg: [] }],
+                mock: [{ spy: vi.spyOn(cmdProviders.mvs, "issueConsoleCommand"), arg: [] }],
             },
             {
                 name: "zowe.selectForCompare",
@@ -368,7 +368,7 @@ describe("Test src/shared/extension", () => {
 
         let zosmfRestClientSetThrottleOptionsSpy: MockInstance;
         beforeAll(() => {
-            vi.spyOn(MvsCommandHandler, "getInstance").mockReturnValue(cmdProviders.mvs as any);
+            vi.spyOn(ConsoleCommandHandler, "getInstance").mockReturnValue(cmdProviders.mvs as any);
             vi.spyOn(TsoCommandHandler, "getInstance").mockReturnValue(cmdProviders.tso as any);
             vi.spyOn(UnixCommandHandler, "getInstance").mockReturnValue(cmdProviders.uss as any);
             vi.spyOn(SettingsConfig, "getDirectValue").mockImplementation((key: string, defaultValue?: unknown): unknown => {
@@ -478,7 +478,7 @@ describe("Test src/shared/extension", () => {
                 writable: true,
             });
 
-            vi.spyOn(MvsCommandHandler, "getInstance").mockReturnValue({ issueMvsCommand: vi.fn() } as any);
+            vi.spyOn(ConsoleCommandHandler, "getInstance").mockReturnValue({ issueMvsCommand: vi.fn() } as any);
             vi.spyOn(TsoCommandHandler, "getInstance").mockReturnValue({ issueTsoCommand: vi.fn() } as any);
             vi.spyOn(UnixCommandHandler, "getInstance").mockReturnValue({ issueUnixCommand: vi.fn() } as any);